Update the Ride class according to Part 3
In the second part of the exercise, I have created an interface for the Ride and have initially created eight methods. In the third part, I refined the first three methods on queuing, updated the Ride class with queues, and made sure that all of them printed appropriate prompts. Finally, I created the object as required and called the method to demonstrate the functionality of the code.