Codes for Parts 4A and 4B
According to the requirements of Part 4A, the four interface methods of the Ride class have been updated to make it more complete and print out a reminder message every time. According to Part 4B, a new Comparator class is created to implement the Compare method, which can sort visitors.