Update the code for Part 5
I updated the method of running a ride according to the requirements of Part 5, set the maximum number of people that can be carried per run and the total number of runs of the device, and also sent a message warning for some problems, such as the failure to run the ride when the queue is empty. Finally, I created the corresponding object and called the method to prove the functionality of the code.