Update the code for Parts 6 and 7
In the fifth module, you learned how to import and export data to a file. Parts VI and VII deal with the handling of documents. In these two sections, I created two methods for exporting data and importing data. In Part 6, I first created a new Ride and some Visitor objects, then exported the visitor as required into the file I created, and then in Part 7 I imported and exported the previously created file. For exception handling, I make sure to add appropriate error message alerts in both methods. In addition, I also created a printfile method to print the information stored in the file, which is more intuitive.